requirements.txt
streamlit
azure-cognitiveservices-speech
openai
numpy
pandas
matplotlib
fpdf
python-dotenv
asyncio
aiohttp